UIMA-2313 fixed SETFEATURE, FEATURE, GETFEATURE, MARKTABLE and actions based on structure (CREATE) for other primitive types. Tested with floats and SETFEATURE. Fixed wrong return type in number expressions.
git-svn-id: https://svn.apache.org/repos/asf/uima/sandbox/trunk/TextMarker@1229297 13f79535-47bb-0310-9956-ffa450edef68

+++ uimaj-ep-textmarker-engine/src/main/java/org/apache/uima/textmarker/expression/number/ComposedNumberExpression.java
@@ -54,14 +54,14 @@ public class ComposedNumberExpression extends AbstractNumberExpression {
   }
   
   @Override
-  public double getFloatValue(TextMarkerStatement parent) {
+  public float getFloatValue(TextMarkerStatement parent) {
     NumberExpression numberExpression = getExpressions().get(0);
     if (numberExpression == null) {
       return 0;
     }
-    double result = numberExpression.getFloatValue(parent);
+    float result = numberExpression.getFloatValue(parent);
     for (int i = 0; i < getOperators().size(); i++) {
-      double second = 0;
+      float second = 0;
       if (getExpressions().size() > i + 1) {
         second = getExpressions().get(i + 1).getFloatValue(parent);
       }
